evenodds fix

+++ codeforces/evenodds.cpp
@@ -4,10 +4,10 @@
 using namespace std;
 
 int main() {
-    int n, k;
+    long long n, k;
     cin >> n >> k;
 
-    int odd = (n+1)/2;
+    long long odd = (n+1)/2;
 
     if (k <= odd) {
         cout << (k*2) - 1 << endl;
